requirements.txt

usd-core
numpy
pillow
